+++ www/modules/jitsimeet/JitsimeetModule.php
@@ -42,9 +42,34 @@ class JitsimeetModule extends \GO\Base\Module{
 //		}
 //	}
 
+	private static function customBase64($string): string {
+		return str_replace(['+', '/', '='], ['-', '_', ''], base64_encode($string));
+	}
+	private static function createJwtToken($appIs, $room, $secret): string {
+		$header = json_encode(['typ' => 'JWT', 'alg' => 'HS256']);
+		$payload = json_encode(['aud' => $appIs, 'iss' => $appIs, 'room' => $room]);
+		
+		$base64UrlHeader = self::customBase64($header);
+		$base64UrlPayload = self::customBase64($payload);
+		
+		$signature = hash_hmac('sha256', $base64UrlHeader . "." . $base64UrlPayload, $secret, true);
+		$base64UrlSignature = self::customBase64($signature);
+		
+		return $base64UrlHeader . "." . $base64UrlPayload . "." . $base64UrlSignature;
+	}
+	
+	static function generateRoom(): string {
+		$settings = Settings::get();
+		$appId = $settings->jitsiJwtAppId;
+		$room =  StringUtil::random(5);
+		return $settings->jitsiJwtEnabled ? $room .'?jwt=' .self::createJwtToken($appId, $room, $settings->jitsiJwtSecret) : $room;
+	}
 	public static function checkIfHasLink($self, &$response,$model,&$params){
 		//find link in description
-		$response['data']['jitsiMeet'] = self::hasLink($model);
+		$hasLink = self::hasLink($model);
+		$response['data']['jitsiMeet'] = $hasLink;
+		if(!$hasLink)
+			$response['data']['jitsiRoom'] = self::generateRoom();
 	}
 
 	public function getSettings() {